requirements.txt
absl-py==2.1.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==24.2.0
blinker==1.6.3
bottleneck==1.3.5
branca==0.7.2
cachetools==5.5.0
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cryptography==39.0.1
cycler==0.11.0
et_xmlfile==1.1.0
folium==0.16.0
fonttools==4.38.0
geopandas==0.10.2
google-auth==2.35.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.10
importlib-metadata==4.11.4
jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.4
markdown
markupsafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
openpyxl==3.0.10
osmnx==1.1.2
packaging==23.2
pandas==1.3.5
pillow==9.2.0
protobuf==4.21.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyjwt==2.8.0
pyopenssl==23.2.0
pyparsing==3.1.2
pyproj==3.2.1
pysocks==1.7.1
python-dateutil==2.9.0
pytz==2024.2
pyu2f==0.1.5
requests
requests-oauthlib==2.0.0
rsa==4.9
rtree==1.0.1
scikit-learn==0.22.2.post1
scipy==1.7.3
shapely==1.8.2
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
typing-extensions==4.7.1
werkzeug==2.2.3
wheel==0.42.0
zipp==3.15.0
